Fix errors in rumble data populating
When $scope.rumble is set to an empty object, it does not contain the
properties "ability" and "special", so denormalizeEffects() receives
undefined. One way would be to add these properties to the empty object.
Another would be to use conditions before calling the function. I
decided to just check it inside the function as it makes the least
changes.

Also fixed a typo that should have been "rumble2"