Use one-time binding for CharTable sidebar.
Static data do not need watchers (which get processed every digest
cycle). I applied the one-time binding feature for parts in the code
where the data is not expected to change during runtime, namely the
text of the filters (from matchers.js, which is also not changed during
runtime). This is done by adding "::" before the variables. This
removes watchers from the variables after the value is defined
(basically evaluate-once).

I used the Chrome extension "Angular Watchers" from
chrome.google.com/webstore/detail/angular-watchers
/nlmjblobloedpmkmmckeehnbfalnjnjk
to test out the number of watchers in the page. Once it is installed,
it can be found in the browser's developer tools as a tab named
"$$Watchers" (unfortunately not mentioned in the
extension's description)

I made sure that each change reduces the number of watchers, while not
affecting functionality (turning the filters on and off are working).
The observed watcher count reduction for each type of change
for #captain-container and the succeeding elements is listed as follows.

Variables (what appears as text): Number of filters for the section
    (47 for captain, 113 for special, etc)
ngRepeat: 1
ngIf: 279 (number of all matchers)

Interestingly, ngIf has watchers for each element in ngRepeat's array,
so basically each section (captain, special, support, etc) had watchers
for all of them, which is 279 (as of this writing) for EACH section.

Note that I did not use it for the ngClass directives, as we do need
those bindings for appearance.

Applying this to all sections from captain ability filters reduced
watchers from 2857 to 899 (~69% reduction).

I then applied this to all the other sections with ng-Repeat.
Even if the arrays in their ngRepeat directives were literal arrays,
their variables still had a few watchers. This brought the final count
to 857.

Tested in Microsoft Edge 91.0.864.48 (64-bit).

I do hope to apply this feature in other views as well.

Additional references:
https://docs.angularjs.org/guide/expression#one-time-binding
https://www.qualtrics.com/eng/tuning-angularjs-performance/