Make searching independent of word order
This puts all the searchTerms in an array so that they may be
tested individually. This will reduce the need for multiple variants of
aliases. Like the prior code, the searchTerms are matched against a
string composed of the unit name and the aliases. The previous
`result.query` is retained for the fuzzy search, though it is now
returned as a string instead of a regex. If the fuzzy search can be
revised to use `result.queryTerms`, `result.query` may be removed.

String and mathematical operators in the query are unaffected, and can
be in between searchTerms.

Simple tests to see the difference between the previous searching and
this commit's are "legend v2 akainu", "nami hug", and "tm zanji".