Add family and notfamily operator
This lets you search for specific unit families, thus preventing false
positives such as Boa appearing when you search for Luffy. This allows
accurate filtering for supported characters. Syntax is `family:Luffy`.
Note that this is still a partial search, so the query could match other
families containing the word, such as 'Arlong' matching
'Arlong Crewmember'. As such, to make the query more accurate, you can
use the regex special characters to denote a full-match like
`family:^Arlong$` and `family:^Monkey_D._Luffy$`. Note that instead of
spaces, you need to use underscores as the query is split by spaces.

This also adds the families to the default search.

As this converts the query (after the `family:` part) into regex, one
must take care when using regex special characters in the search, like
'(', '[', and '.'. If they are a part of a family name (which should be
avoided to prevent confusion, except for the dot), use a backslash in
front of each special character to escape them (make them regular
characters) like `\[`.

The `notfamily` operator allows excluding a family from the results.
Units with multiple families that have the given family name will be
excluded. This can be for excluding the supporting unit when searching
for supported characters.