Revamp families
This change will allow these families to have family-based aliases, so
users can search with the family operator `family:X` with multiple
options, such as "Cutty Flam" and "Franky", while also making it simple
to add more in the future. This also lets us add aliases that should
apply to similar units. The variable names are in order of appearance
in the game (unit ID of first appearance of the family).

Families for ghost units (dual and VS units) are added so that they may
be searchable in the damage calculator.

The hack job for Karoo being a socket for some Vivis is fixed by adding
a karoo family. The same goes for Demalo Black.

The spread operator is used in window.families to keep the values in
window.families 1D arrays, thus keeping the structure simple. The words
"and" and "or" are not allowed in the family names as they are keywords
for parsing supported characters and super special criteria. You may use
` & ` as the separator instead, or split them into two families.

Do not add partial family aliases because it will not change anything,
other than add another string. For example, you do not have to add
"Luffy" for "Monkey D. Luffy" as the family operator will search
partially by default.

The logic for getting the drop locations of farmable versions is
refactored and now uses a reverseFamilyMap, improving performance by
around 50% for the mentioned block, which otherwise would easily become
slower as family aliases are added. The other slow part in the block is
CharUtils.searchDropLocations, which can be optimized later on by adding
a reverse map for it too, or modifying reverseDropMap to include the
necessary data.