Refactor Sugar toys and Toki-toki state
They will now be stored directly into the `data.team` array, instead of
having 6 variable each. Because of this change, moving a unit that is a
sugar toy or has tokiState will now keep the state along with the unit,
instead of locking the state to a team position. `tokiState` was used
instead of just `toki` since the former is more readable (the
codebase already has many confusing variable names).

Also, labels were added in sidebar.html to follow best practices (also
allows clicking on the label to toggle).