Refactor(damage): Reduce repeated code
Ternary if chains were simplified (they're very confusing and really
just makes it harder to read the code). If ternary if expressions are
simple enough, like one condition and two values, it's fine.

Recomputed functions were stored in variables (especially ones with
`getParameters()`, which should be fine as it does not seem like any
function actually modifies the resulting params, at least in the ones I
touched). Not only does this reduce the code, but it also reduces
function calls.

'1.25x Chain Multiplier - Sanji Judge Change Action' effect was removed
in the calc as it seems that it is no longer necessary. Was kept in
effect.js as it seems to be forbidden to remove anything there.