Feat: Add supports operator
Allow filtering for units that can support the specified unit.
`supports:1` will show units that can support unit 1.

This recreates the query generated for each unit with a support to check
the unit being queried for, which, admittedly, is not very quick. To cut
down on the calls, this makes `Utils.generateCriteriaQuery` return an
object similar to what `Utils.generateSearchParameters` returns, so that
we don't have to parse the query again (which was my initial solution).

This also allows a query for the supports for multiple characters, like
`supports:1|5`, though I'm not sure if it is actually useful to do so.