Fix: Use relative path for thumbnails
Different thumbnail urls were used for setups where the root folder
of the db is different from the root directory of the server because of
the absolute path. This is not an issue for the live site, because the
url is simply "https://optc-db.github.io/api/images", which is where
"/api/images" points to. However, forks seem to have a different url.
My url is "https://codeyan01.github.io/optc-db.github.io/api/images"