Optimize datatable rendering
`deferRender: true` improves performance significantly as HTML elements
will only be created when needed (even if only 10 are shown, it would
try to create one for all rows at initialization if the option is off).
See: https://datatables.net/reference/option/deferRender
Also: https://datatables.net/forums/discussion/46654/why-the-deferrender-default-value-is-set-as-false

Only caveat is that it isn't guaranteed to have all nodes when using
DataTables API methods that target all nodes.

The column.render function `addImage` would be called for 4 datatable
events: display (rendering of the element), sorting, filtering, and type
(no idea what 'type' is). As such, the <img> element should only be
returned when the event is of the "display" type, and the normal string
for the rest. This change makes it so Utils.getThumbnailUrl only gets
called once for each row during initialization