Optimize Details View
Use one-time binding for all bindings in details.html, except for those
ones that are used for comparing between two units (since you need them
to update).

For addTags and addNames, append all html string to one string first,
then append once. This makes it so the html is compiled only once, not
every time a tag is appended.