Fix supports operator bug with incomplete units
When searching with `supports:X` where the unit has no families yet, or
the unit does not exist, the table does not update to show feedback.
The fix is to check if unit X has families. If it does not, we continue
evaluating the unit against other supports (like type/class/cost/all
characters supports).