Ignore class/type condition for super special
This prioritizes the family names condition instead. This stops the test
for super special queries to keep failing on unit 3609 Kaido. This is
however just a workaround.

idea if someone's gonna work on it: return an array of queries, so that
there will be multiple "Search for these characters" links, depending on
the conditions (so that one will handle the `6 Striker characters`,
the other will handle the family names.)