Add immunities and cached params
For `ignoreImmunities`, it has to be a function that returns an array of
strings of the enemyImmunities (see app.js, $scope.data.enemyImmunities)

`p.cached` allows you to access the cached params object. So
`p.cached.captain` gives you the captain at the time of activation, and
`p.captain` gives you the captain at the time of execution of the
function. This eliminates the need to implement nearly unreadable code
with `onActivation`. Since the cached params are only stored in
`p.cached`, this caching will not affect existing code that does not use
it.

Caching is done on `specialToggled` and `altspecialToggled` in
cruncher.js. It is important that the cached parameter objects for the
unit's special and its altspecial are different, which is why
`cacheParameters` requires the special type to be passed (which is used
as a key). It also require the `sourceSlot` (index of the unit that
launched the special), since the cached objects should be tied to the
unit that launched it.

Thus, `getParameters` is modified to accept `sourceSlot`, which also
eliminates the need to manually insert the sourceSlot property most of
the time. While this property is also used to retrieve the cached
parameters, some ability functions do not require caching (captains,
sailors, etc), so it is fine to pass the `sourceSlot` without the
`specialType`.

The enemy effects are refactored to bundle `isDelayed`, `isPoisoned`,
etc. together, for better context, to make it universal
(isIncreasedDamageTaken imo doesn't look good), and to mirror the new
`enemyImmunities` object, making it easier to implement the immunities,
since the object keys will be the same.

I wish to phase out the use of `p.delayed`, `p.poisoned`, etc., in
favor of the above.

Enemy effects applied by specials are now evaluated upon toggling the
special, so that toggling an immunity after a special should not remove
the enemy effect.

This commit changes the existing specials that check "If enemies have
Delay Debuff Protection or All Debuff Protection at the time the special
was activated" to no longer assume that the enemy has the immunities.
Since the enemy could have no immunity upon launching the first unit,
but have immunity upon launching the second unit (e.g. the special from
the first unit is extended into the next stage), checks for immunities
use `p.cached.enemyImmunities.delay` (or whatever immunity). While this
commit also modified some specials to use the `cached` property
instead of the onActivation pattern, this does not touch all specials,
as that is too many. It is, however, advised to clean up the other
specials later on.

All existing increasedDamageTaken (IDT) applying units have also been
implemented, except for 3609 Kaido's CA that applies IDT upon tapping
the unit with a wano slot, since the existing code does not allow to
implement that trivially.

IDT is applied after damage with defense is calculated, and then rounded
down. Hence, it is implemented after chain boosts.

Also adds a dummy captain and friend captain unit, so that code in
`specials.js` and related files do not need to do
`p.captain && p.captain.type...` in case there is no captain.

Not all specials were modified to use `p.cached.multiplier` instead of
`onActivation`, since there are already too many, but I did a few just
to make examples. Those that use multiple stages also were not touched.
window.specials[...].multiplier were changed to `p.cached.multiplier`
though. Note that only properties that contain "multiplier" were
replaced with this form (such as "orbmultiplier" and "multiplier2").

transfer.js was modified to only enable specials when the cruncher is
actually ready to process `onActivation` calls. previously it would
first enable the specials, which would let the cruncher already use the
specials without first calling `onActivation` methods. In addition, it
used to emit `specialToggled` even for disabled specials, which is wrong
since `onDeactivation` functions were called without even enabling the
special.